Adjusted tcaselect and how channels are referred to in messages
Updated TCA_ADDRESS to be a variable as opposed to a table since we have no plans of running multiple multiplexers on the same autopilot. Changed tcaselect to reflect this change so that is simply uses the TCA_ADDRESS variable as opposed to the user passing in the index of the TCA in the array.

Changed alot of the wording in the messages that get sent to the ground control station and to the BIN file to be a bit more clear and consistent with how the sensors are labled on the board. This mainly changes us from reporting the channel and its number on the multiplexer, to us now calling each sensor a 'tube
' since they're connected to tubes, and not indexing at 0, index at 1.