Made comments more clear and added a new function to dynamically create status messages
Added and edited comments to either add more clarity to parts of the
code, or adjust out of date comments from earlier revisions.

Added the function 'form_message' to dynamically create the status
message that gets sent to Mission Planner. This prevents us from having
to make several edits to the message when we want to edit the
formattings, or prevent us from having to edit the message when we
change the number of channels we are reading from.