Changed name of i2c_tca_2.lua
Changed the name of i2c_tca_2.lua to i2c_5hole_probe.lua to improve the
clarity of the name, since it was a carry over of the i2c_tca.lua name,
which now does not best describe the purpose of the script. Especially
for those not already aware of what the script does.