Serial Anemometer: Changed how data is pulled from serial line
Changed the primary loop in the update function to use the readstring()
method for the serial interface instead of reading bytes one at a time
from the serial line.

Reading one byte at a time is two slow for the anemometer since it sends
4 messages at 2Hz. We read the longest possible message length for the
anemometer from the serial line, and parse it as normal. The loop is set
up to run fast enough to where we pull only when there is one message in
the queue. If we read parts of two messages, like when we first boot, we
simply clear the queue to "realign" the messages.

Added better comments for function along with LuaDoc compatible tags for
parameters and return values.