Samamet: Redid update() and added better frame verification
Redid the primary loop in the update function to utilize the
readstring() method to read the entire 60 byte message without having to
read one byte from the serial line at a time. This reduces a lot of the
added complexity that was needed to get the message string consistently.

Since the new method does remove some of the checks that were "built in"
to the previous method (mainly some basic message frame verification
that was done by reading the line until we find a <LF> character), I
added the verify_valid_frame() to check if the message we have read from
the serial line is a valid NMEA-0183 style message, (message starts with
"$" and ends with <CR><LF).

Added some constants (SCHEDULE_RATE and TIME_BETWEEN_DATA) to calculate
the number of "loops to fail" before we log a disconnected sensor error
at run time. Allows us to change the schedule rate without changing all
return statements in update by hand.