Anemometer: Updated error messages and removed references to un-parsed message
Adjusted the ERROR_LIST table to make sure that the error messages are
16 bytes long or shorter, since the data type we use to log the errors
has a max of 16 bytes

Also added the word "ANEM" in the error messages that are sent to
Mission Planner to be clear when running both the Anemometer and Samamet
since they both have similar messages.

Removed some of the calls to the wind temperature message since we are
currently not logging it.

Added some better comments to functions and other parts of the code.