Samamet: Improved extraction of data from messages
Added the table MESSAGE_INFO with the message header of a message as the
key, and a table which holds the length of the message, the number of
fields in the message, and the number of measurements in the message.

This is used in parse_data and log_data, where they verify that the
number of elements in the data_table and measurement_table matches with
the number of fields and measurements that we expect for the message we
are parsing.

The table can be expanded to hold more messages, but will not be needed
in this case, but can be useful for decoding messages from other
NMEA-0183 devices.

Also changed how parse_data extracts measurements to use pattern
matching to extract the numbers from the fields of the message.

This is very similar to the changes to the Anemometer script in commit
3f07f830cd60eb9f374df3921ec8bb812adb3765.