Anemometer: more formatting changes to inline verbatim
Org-mode by default does not allow for certain characters to be inside
inline code, verbatim, and other objects. You can change this locally so
it is not an issue, but does not work with GitHub's org-mode rendering
since it is separate.

To prevent one of the verbatim blocks containing a regex to be
improperly rendered, I finally broke over and placed it in an example
block.

Note: you can get around this sometimes by adding the zero width space
unicode character, "U+200B" before the problem characters. However this
can make editing the document harder in the future since the character
is invisible in any text editor that supports UFT-8 (bascially
everything).