Anem & Sama: Updated first return to clear serial queue
Added a statement to both the Samamet and anemometer scripts to clear
the queue after the initialization process to ensure that we have a
clean queue to process once the update function is scheduled.

This primarily effects the anemometer since it sends more messages more
often. Without this statement, the anemometer script would send an error
on startup since there would be more than one message in the queue. The
script would fix itself of course but there is no reason to not remove
an easily preventable error message.