Multi Hole Probes: Added variable GCS report rates
Added the ability to adjust the rate at which the script reports data to
the ground control stations (Mission Planner in our case). The script
will only send text if it has been REPORT_RATE or more milliseconds
since the last transmission.

Also adjusted the order of the constants to be more logical.